Move FMatrix3x3 to separate file
Boy scouting this, since I'm about to add some functionality to this class. According to our code style, every class must be in its own file, and the implementation must be in the cpp files.

Contributes to issue CURA-7118.
