+++ CMakeLists.txt
@@ -196,6 +196,7 @@ set(engine_SRCS # Except main.cpp.
     src/utils/AABB.cpp
     src/utils/AABB3D.cpp
     src/utils/Date.cpp
+	src/utils/FMatrix3x3.cpp
     src/utils/gettime.cpp
     src/utils/getpath.cpp
     src/utils/LinearAlg2D.cpp
