+++ src/communication/ArcusCommunicationPrivate.cpp
@@ -1,4 +1,4 @@
-//Copyright (c) 2018 Ultimaker B.V.
+//Copyright (c) 2020 Ultimaker B.V.
 //CuraEngine is released under the terms of the AGPLv3 or higher.
 
 #ifdef ARCUS
@@ -8,7 +8,8 @@
 #include "../ExtruderTrain.h"
 #include "../Slice.h"
 #include "../settings/types/LayerIndex.h"
-#include "../utils/floatpoint.h"
+#include "../utils/floatpoint.h" //To accept vertices (which are provided in floating point).
+#include "../utils/FMatrix3x3.h" //To convert vertices to integer-points.
 #include "../utils/logoutput.h"
 
 namespace cura
