+++ src/utils/floatpoint.h
@@ -1,4 +1,4 @@
-//Copyright (c) 2018 Ultimaker B.V.
+//Copyright (c) 2020 Ultimaker B.V.
 //CuraEngine is released under the terms of the AGPLv3 or higher.
 
 #ifndef FLOAT_POINT_H
@@ -100,32 +100,7 @@ inline float operator*(FPoint3 lhs, const FPoint3& rhs) {
 //  return lhs;
 //}
 
-class FMatrix3x3
-{
-public:
-    double m[3][3];
 
-    FMatrix3x3()
-    {
-        m[0][0] = 1.0;
-        m[1][0] = 0.0;
-        m[2][0] = 0.0;
-        m[0][1] = 0.0;
-        m[1][1] = 1.0;
-        m[2][1] = 0.0;
-        m[0][2] = 0.0;
-        m[1][2] = 0.0;
-        m[2][2] = 1.0;
-    }
-    
-    Point3 apply(const FPoint3& p) const
-    {
-        return Point3(
-            MM2INT(p.x * m[0][0] + p.y * m[1][0] + p.z * m[2][0]),
-            MM2INT(p.x * m[0][1] + p.y * m[1][1] + p.z * m[2][1]),
-            MM2INT(p.x * m[0][2] + p.y * m[1][2] + p.z * m[2][2]));
-    }
-};
 
 }//namespace cura
 #endif//INT_POINT_H
