+++ tests/integration/SlicePhaseTest.cpp
@@ -1,4 +1,4 @@
-//Copyright (c) 2019 Ultimaker B.V.
+//Copyright (c) 2020 Ultimaker B.V.
 //CuraEngine is released under the terms of the AGPLv3 or higher.
 
 #include <gtest/gtest.h>
@@ -6,7 +6,7 @@
 #include "../src/Application.h" //To set up a slice with settings.
 #include "../src/Slice.h" //To set up a scene to slice.
 #include "../src/slicer.h" //Starts the slicing phase that we want to test.
-#include "../src/utils/floatpoint.h" //For FMatrix3x3 to load STL files.
+#include "../src/utils/FMatrix3x3.h" //To load STL files.
 #include "../src/utils/polygon.h" //Creating polygons to compare to sliced layers.
 #include "../src/utils/polygonUtils.h" //Comparing similarity of polygons.
